requirements.txt
annotated-types==0.7.0
anyio==4.5.2
asgiref==3.8.1
backports.zoneinfo==0.2.1
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
cryptography==45.0.3
deprecation==2.1.0
dj-database-url==2.3.0
Django==4.2.19
django-cors-headers==4.4.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
ecdsa==0.19.1
exceptiongroup==1.3.0
Faker==35.2.2
gotrue==1.3.1
h11==0.14.0
httpcore==0.17.3
httpx==0.24.1
idna==3.10
packaging==25.0
postgrest==0.11.0
psycopg2-binary==2.9.10
pyasn1==0.4.8
pycparser==2.22
pydantic==2.10.6
pydantic-core==2.27.2
PyJWT==2.9.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-jose==3.4.0
realtime==1.0.6
requests==2.31.0
rsa==4.9.1
six==1.17.0
sniffio==1.3.1
sqlparse==0.5.3
storage3==0.6.1
StrEnum==0.4.15
supabase==1.2.0
supafunc==0.2.4
typing-extensions==4.12.2
urllib3==2.2.3
websockets==12.0
